# docs/contratos/api/admin-v1.openapi.yaml
openapi: 3.1.0
info:
  title: ModelCairn Administration API
  version: 1.0.0-alpha.1
  description: |
    Draft Milestone 3 contract. Administrative responses use Cache-Control: no-store.
    Login and mutations require the configured exact Origin; online CLI clients
    supply it too. Authenticated operations require session and CSRF except
    session/me, which requires the same-origin recovery checks in admin-runtime-v1.
    Authentication failures return 401; Origin/CSRF failures return 403 before
    business mutations. These boundary errors apply to all protected operations.
servers:
  - url: /api/v1/admin
security:
  - adminSession: []
    csrfToken: []
paths:
  /session:
    post:
      security: []
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Login' }
      responses:
        '200':
          description: Session created; token is also returned by session/me after reload
          headers:
            Set-Cookie:
              schema: { type: string }
              description: mc_session with HttpOnly, SameSite=Strict, Path=/ and Secure on HTTPS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionContext' }
        '401': { $ref: '#/components/responses/Error' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      operationId: deleteSession
      responses:
        '204': { description: Session revoked and cookie cleared }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /overview:
    get:
      operationId: getOverview
      description: Bounded content-free operational summary for the console landing page.
      responses:
        '200':
          description: Resource totals, last-24-hour outcomes, active cooldown count, and at most five recent requests.
          content: { application/json: { schema: { $ref: '#/components/schemas/Overview' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /requests:
    get:
      operationId: listOperationalRequests
      description: Content-free operational history ordered by start time and immutable request id.
      parameters:
        - { name: outcome, in: query, schema: { type: string, enum: [success, error, partial, cancelled, indeterminate] } }
        - { name: alias, in: query, schema: { type: string, maxLength: 128 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200': { description: Bounded page, content: { application/json: { schema: { $ref: '#/components/schemas/OperationalRequestPage' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /requests/{id}/attempts:
    get:
      operationId: listOperationalAttempts
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, maxLength: 128 } }
      responses:
        '200':
          description: At most 32 attempts ordered by sequence.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties: { items: { type: array, maxItems: 32, items: { $ref: '#/components/schemas/OperationalAttempt' } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
  /resources/{kind}:
    parameters:
      - $ref: '#/components/parameters/Kind'
    get:
      operationId: listResources
      parameters:
        - { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200':
          description: Page ordered by immutable uid
          content: { application/json: { schema: { $ref: '#/components/schemas/ResourcePage' } } }
        '400': { $ref: '#/components/responses/Error' }
    post:
      operationId: createResource
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } }
      responses:
        '201': { $ref: '#/components/responses/Resource' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /resources/{kind}/{name}:
    parameters:
      - $ref: '#/components/parameters/Kind'
      - $ref: '#/components/parameters/Name'
    get:
      operationId: getResource
      responses:
        '200': { $ref: '#/components/responses/Resource' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      operationId: replaceResource
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } }
      responses:
        '200': { $ref: '#/components/responses/Resource' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteResource
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
  /config/validate:
    post:
      operationId: validateConfiguration
      requestBody: { $ref: '#/components/requestBodies/Configuration' }
      responses:
        '200': { $ref: '#/components/responses/Validation' }
        '400': { $ref: '#/components/responses/Error' }
  /config/plan:
    post:
      operationId: planConfiguration
      parameters:
        - { name: allowDelete, in: query, schema: { type: boolean, default: false } }
      requestBody: { $ref: '#/components/requestBodies/Configuration' }
      responses:
        '200': { $ref: '#/components/responses/Plan' }
        '400': { $ref: '#/components/responses/Error' }
  /config/apply:
    post:
      operationId: applyConfiguration
      parameters:
        - { name: allowDelete, in: query, schema: { type: boolean, default: false } }
        - $ref: '#/components/parameters/PlanToken'
      requestBody: { $ref: '#/components/requestBodies/Configuration' }
      responses:
        '200': { $ref: '#/components/responses/ApplyResult' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
  /secrets/{name}:
    get:
      operationId: getSecretMetadata
      parameters:
        - $ref: '#/components/parameters/Name'
      responses:
        '200': { $ref: '#/components/responses/SecretMetadata' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      operationId: setSecret
      parameters:
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/IfMatchOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [value]
              properties:
                value:
                  type: string
                  minLength: 1
                  maxLength: 16384
                  writeOnly: true
                  description: Server enforces 8–16384 UTF-8 bytes; JSON Schema length counts characters.
      responses:
        '200': { $ref: '#/components/responses/SecretMetadata' }
        '201': { $ref: '#/components/responses/SecretMetadata' }
        '400': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deleteSecret
      parameters:
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
  /agent-tokens/{name}/issue:
    post:
      operationId: issueAgentToken
      parameters: [ { $ref: '#/components/parameters/Name' } ]
      responses:
        '201':
          description: Token value returned exactly once
          content:
            application/json:
              schema:
                type: object
                required: [token, tokenStatus]
                additionalProperties: false
                properties:
                  token: { type: string, readOnly: true, pattern: '^mc_at_v1_[A-Za-z0-9_-]{43}$', description: One-time intentional delivery to the authenticated issuer; never logged or recoverable later. }
                  tokenStatus: { $ref: '#/components/schemas/AgentTokenStatus' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /agent-tokens/{name}/revoke:
    post:
      operationId: revokeAgentToken
      parameters: [ { $ref: '#/components/parameters/Name' } ]
      responses:
        '204': { description: Revoked immediately for new requests }
        '404': { $ref: '#/components/responses/Error' }
  /agent-tokens/{name}/status:
    get:
      operationId: getAgentTokenStatus
      parameters: [ { $ref: '#/components/parameters/Name' } ]
      responses:
        '200':
          description: Safe credential lifecycle metadata; never a bearer or verifier
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [tokenStatus]
                properties: { tokenStatus: { $ref: '#/components/schemas/AgentTokenStatus' } }
        '404': { $ref: '#/components/responses/Error' }
  /strategies/{name}/publish:
    post:
      operationId: publishStrategy
      parameters:
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200': { $ref: '#/components/responses/Resource' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '412': { $ref: '#/components/responses/Error' }
        '428': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    adminSession: { type: apiKey, in: cookie, name: mc_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Kind:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [providers, provider-accounts, provider-connections, credentials, egresses, models, destinations, strategies, routes, agent-tokens]
    Name: { name: name, in: path, required: true, schema: { type: string, pattern: '^[a-z][a-z0-9-]{0,62}$' } }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Single strong quoted positive resource version. Missing returns 428, malformed (including wildcard or lists) 400, stale 412.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    IfMatchOptional:
      name: If-Match
      in: header
      required: false
      description: Omit only when creating a new secret; replacing an existing secret requires its current strong ETag.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    PlanToken:
      name: X-ModelCairn-Plan-Token
      in: header
      required: true
      description: Opaque token returned by the matching plan endpoint; binds operation purpose, normalized desired document, observed revision and expiry. Configuration and settings tokens are not interchangeable.
      schema: { type: string, minLength: 32 }
  requestBodies:
...
    Configuration:
      required: true
      content:
        application/json:
          schema: { $ref: '../config/modelcairn-config-v1alpha1.schema.json' }
        application/yaml:
          schema: { $ref: '../config/modelcairn-config-v1alpha1.schema.json' }
  responses:
    RateLimited:
      description: Login admission denied; retry after the indicated number of seconds.
      headers:
        Retry-After:
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error:
      description: Structured error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Resource:
      description: Resource with ETag header
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Resource' } } }
    SecretMetadata:
      description: Metadata only; never the secret value
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/SecretMetadata' } } }
    Validation:
      description: Validation result
      content: { application/json: { schema: { $ref: '#/components/schemas/Validation' } } }
    Plan:
      description: Redacted ordered changes
      content: { application/json: { schema: { $ref: '#/components/schemas/Plan' } } }
    ApplyResult:
      description: Redacted result of the atomic apply
      content: { application/json: { schema: { $ref: '#/components/schemas/ApplyResult' } } }
  schemas:
...
    Login:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 120 }
        password: { type: string, minLength: 12, writeOnly: true, description: Maximum 1024 UTF-8 bytes enforced by server }
    Admin:
      type: object
      required: [id, username]
      properties: { id: { type: string, format: uuid }, username: { type: string } }
    SessionContext:
      type: object
      required: [admin, csrfToken, expiresAt]
      properties:
        admin: { $ref: '#/components/schemas/Admin' }
        csrfToken: { type: string, description: Kept in SPA memory and sent as X-CSRF-Token }
        expiresAt: { type: string, format: date-time }
    Overview:
      type: object
      additionalProperties: false
      required: [resourceCounts, requests24h, activeCooldowns, recentRequests, generatedAt]
      properties:
        resourceCounts:
          type: object
          additionalProperties: { type: integer, minimum: 0 }
        requests24h:
          type: object
          additionalProperties: false
          required: [total, success, error]
          properties:
            total: { type: integer, minimum: 0 }
            success: { type: integer, minimum: 0 }
            error: { type: integer, minimum: 0 }
        activeCooldowns: { type: integer, minimum: 0 }
        recentRequests:
          type: array
          maxItems: 5
          items:
            type: object
            additionalProperties: false
            required: [id, requestedAlias, startedAt, attempts]
            properties:
              id: { type: string, format: uuid }
              requestedAlias: { type: string }
              startedAt: { type: string, format: date-time }
              completedAt: { type: [string, 'null'], format: date-time }
              outcome: { type: [string, 'null'], enum: [success, error, partial, cancelled, indeterminate, null] }
              httpStatus: { type: [integer, 'null'], minimum: 100, maximum: 599 }
              durationMs: { type: [integer, 'null'], minimum: 0 }
              attempts: { type: integer, minimum: 0 }
        generatedAt: { type: string, format: date-time }
    OperationalRequestPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/OperationalRequest' } }
        nextCursor: { type: [string, 'null'], maxLength: 512 }
    OperationalRequest:
      type: object
      additionalProperties: false
      required: [id, requestedAlias, startedAt, attempts]
      properties:
        id: { type: string, maxLength: 128 }
        requestedAlias: { type: string, maxLength: 128 }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        outcome: { type: [string, 'null'], enum: [success, error, partial, cancelled, indeterminate, null] }
        httpStatus: { type: [integer, 'null'], minimum: 100, maximum: 599 }
        inputTokens: { type: [integer, 'null'], minimum: 0 }
        outputTokens: { type: [integer, 'null'], minimum: 0 }
        ttftMillis: { type: [integer, 'null'], minimum: 0 }
        durationMillis: { type: [integer, 'null'], minimum: 0 }
        attempts: { type: integer, minimum: 0, maximum: 32 }
    OperationalAttempt:
      type: object
      additionalProperties: false
      required: [sequence, startedAt, outcome, retryable]
      properties:
        sequence: { type: integer, minimum: 1, maximum: 32 }
        destinationID: { type: [string, 'null'] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        outcome: { type: string, enum: [success, error, partial, cancelled, indeterminate] }
        errorClass: { type: [string, 'null'] }
        providerStatus: { type: [integer, 'null'], minimum: 100, maximum: 599 }
        providerRequestID: { type: [string, 'null'] }
        retryable: { type: boolean }
        fallbackReason: { type: [string, 'null'] }
    Resource:
      $ref: '../config/modelcairn-config-v1alpha1.schema.json#/$defs/resource'
    ResourcePage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Resource' } }
        nextCursor: { type: [string, 'null'] }
...
    AgentTokenStatus:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state: { type: string, enum: [unissued, active, expired, revoked] }
        prefix: { type: string, pattern: '^mc_at_v1_[A-Za-z0-9_-]{8}$' }
        issuedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time }
    Validation:
      type: object
      required: [valid, issues]
      properties:
        valid: { type: boolean }
        issues: { type: array, items: { $ref: '#/components/schemas/Issue' } }
...
    Plan:
      type: object
      required: [changes, valid, planToken, expiresAt]
      properties:
        valid: { type: boolean }
        planToken:
          type: string
          minLength: 32
          description: Opaque short-lived precondition token required by apply; never contains secret values.
        expiresAt: { type: string, format: date-time }
        changes:
          type: array
          items:
            type: object
            required: [operation, kind, name]
            properties:
              operation: { enum: [create, update, noop, delete] }
              kind: { type: string }
              name: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            field: { type: [string, 'null'] }
            retryable: { type: boolean, default: false }